change: character uint to unit64

+++ Web3/test-chain/test/character tests/characterlevelup.test.ts
@@ -2,7 +2,7 @@ import { TestingAppChain } from "@proto-kit/sdk";
 
 import { PrivateKey } from "o1js";
 
-import { UInt32 } from "@proto-kit/library";
+import { UInt64 } from "@proto-kit/library";
 
 import { Character, CharacterKey } from "../../src/character";
 
@@ -36,7 +36,7 @@ describe("Character Level Up Test", () => {
 
         // Create a new character for the key
         const startTX = await appChain.transaction(alice, () => {
-            character.newCharacter(UInt32.from(0));
+            character.newCharacter(UInt64.from(0));
         });
         // Sign the tx
         await startTX.sign();
@@ -47,7 +47,7 @@ describe("Character Level Up Test", () => {
         // Create a new character key
         const aliceCharacter = new CharacterKey({ 
             owner: alice, 
-            id: UInt32.from(1)
+            id: UInt64.from(1)
         });
         // Get the promise
         let startLevelPromise = await appChain.query.runtime.Character.characters.get(aliceCharacter);
@@ -62,7 +62,7 @@ describe("Character Level Up Test", () => {
 
         // Create a tx for testing
         const tx1 = await appChain.transaction(alice, () => {
-            character.levelUP(UInt32.from(1));
+            character.levelUP(UInt64.from(1));
         });
         // Sign the tx
         await tx1.sign();
